Changed refreshAll() in UserData to type double

+++ src/main/java/ca/uwo/csd/cs2212/team09/MainView.java
@@ -61,7 +61,7 @@ public class MainView implements GeneralCallBack {
 	private final static int PAGE_HEART_ZONE = 3;
 	private final static int PAGE_GOALS = 4;
 	
-	private int dailyData[] = {0, 0, 0, 0, 0, 0};
+	private double dailyData[] = {0, 0, 0, 0, 0, 0};
 	private int bestnltDate[] = {0, 0, 0, 0, 0, 0};
 	private String dailyDataMsg[] = {"Calories burned (out)", "Total distance", "Floors climbed", "Steps", "Active minutes", "Sedentary minutes"};
 	public Boolean dailyDataCustomization[] = {true, true, true, true, true, true};
